+++ CHANGES.md
@@ -27,6 +27,7 @@ Core Grammars:
 - fix(swift) ensure keyword attributes highlight correctly [Bradley Mackey][]
 - fix(types) fix interface LanguageDetail > keywords [Patrick Chiu]
 - enh(java) add `goto` to be recognized as a keyword in Java [Alvin Joy][]
+- enh(swift) highlight function and macro call usage [Bradley Mackey][]
 
 New Grammars: